pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wealthpath"
version = "0.1.0"
description = "AI-powered income and wealth projection planner using SCF data"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "numpy>=1.26,<3",
    "pandas>=2,<3",
    "langchain>=0.3,<1",
    "langchain-openai>=0.2,<1",
    "langgraph>=0.2,<1",
    "pyyaml>=6,<7",
    "xgboost>=2.0,<3",
    "shap>=0.45,<1",
    "joblib>=1.3,<2",
    "mlflow>=2.10,<4",
    "pyarrow>=14",
    "scikit-learn>=1.4,<2",
    # FRED MCP integration
    "mcp>=1.0,<2",
    "langchain-mcp-adapters>=0.1,<1",
    "httpx>=0.27,<1",
    # Azure Blob Storage — surrogate model download at startup in production
    "azure-storage-blob>=12.0,<13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "httpx>=0.27,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/wealthpath"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
